Make signatures even more writable
When implementing QtCore.Slot.__signature__, the following
extension to writable signatures became useful:

Signatures were corrected to be writable if they were not set
by the signature module at all. But we should also allow
assignment when the signature was set to None.

This feature is only used for exploration; the correct way to
implement a new signature is to integrate it into the normal
generation process, as being done for QtCore.Slot.__signature__.

Change-Id: I010e505a376993f27c46bd1c14e327e44d375abd
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>