Fix a typing bug in Python 2.7 and update
The typing module has problems if it exists twice.
This gave trouble with matplotlib, when typing was imported
and then called PySide indirectly.

The only reliable cure appears to be to use the typing
module if it is already in sys.modules .

The typing27 modue was updated to the latest version
of https://github.com/python/typing (original file with
additional license headers).

Fixes: PYSIDE-1073
Change-Id: Iebfdfe7fd51080a9334db332719928c001501381
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>