Fix bad shutdown effect on QApplication.instance()
When

	app = QtWidgets.QApplication.instance()

is used before a QApplication has been created, the qApp code
sees no qApp instance in C++ and assumes a shutdown.

This patch keeps track of QApplication being created and
behaves correctly on that aspect.

It is still unsolved that QtCore.QObject.staticMetaObject
gets deleted on a qApp shutdown, which is too much.

I think this can be handled in another patch if at all, since
the shutdown / recreate feature is undocumented and of little use.

Change-Id: I140b6dba45f7cd337580373dbf72bc6d0a625fea
Fixes: PYSIDE-1093
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>