Fix booleans and empty list cases in QSettings
After the fix for PYSIDE-1010 there were two things
that were not properly understood:

1. A special case for booleans was required,
2. When a list was detected, the split was wrongly creating
a list with a '0' instead of an empty one.

Additonally, due to the wrong treatment we couldn't
notice if the value 0 was None, the number zero, or even
false, so this patch amends the previous implementation
to properly treat these cases.

New test cases were added.

Change-Id: I41d5387bb835cfa96f94e5577e993a4b87b303f0
Fixes: PYSIDE-1130
Fixes: PYSIDE-820
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>