Leak the BindingManager singleton.
This is to avoid an infrequent crash-on-exit in Houdini which we suspect
is caused by our non-traditional integration of Python and Qt.  Our
theory is that by the time the BindingManager static object is
destroyed, Qt and Python have shutdown but Python did not destroy all
objects still marked for GC.  Some of those objects are PySide2 objects
which have dangling pointers to destroyed Qt objects.  Therefore, when
the BindingManager attempts to clean up PySide2 references at
destruction time it then produces a seg fault.