Add Python 3.11 support.
Manually cherry-picked a handful of fixes from upstream's
v5.15.12-lts-lgpl tagged commit.  `git cherry-pick` was not
possible because the commits do not belong to any branch
in the upstream repo; they may belong to a fork outside
the repo.

Cherry-picked commits:
- 8b8ebf358e42949159a815c8cc1ce1662fd15b54
  libshiboken: Fix a crash in Shiboken::Object::isValid() for
  Python 3.11
- 5e17aaa3a4fbb4dfaba79fa7ee004dc499cf0add
  Fix usage of Py_TYPE() for Python 3.11
- 1c92e34b1790111e403f265f0faaf4e4d767bd97
  libshiboken: Fix crashes with static strings in Python 3.11
- 39e0f799342b0b1367c96709d71bc9930e24ac6f
  Add support for Python 3.11
- 58b923aae003d5ff39a2cc401078c1b38ba17bc8
  Fix Property GC tracking for Python 3.11
- ae79fb4f865386ce28707e2abb93cb46aa0c3195
  Support Python 3.11.N

These commits are all tagged to issue PYSIDE-1960 but I am
unable to view the issue due to permissions.